Move more data classes to top-level
* Move GitHubRelease to top-level within GitHubReleaseManager.kt
* Move GitHubIssue and GitHubUser to top-level within GitHubSearchManager.kt
* Simplifies other imports, and we can easily move them separate files if
  we ever want/need to
